REFACTOR: view 패키지 추가
뷰 파일을 별도의 패키지로 보관하기 위해 view 패키지를 생성했습니다.
