+++ src/main/java/baseball/view/OutputView.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.view;
 
 public class OutputView {
